Add alternative AST and Types
The ones we already have impose a hierarchy onto blocks, and don't get
rid of blocks when they're destroyed by cuts. My alternative types just
store a bunch of unrelated blocks, and maintains a mapping of who
contains whom; that should be easier to work with, at least for some
algorithms.